test(session-blocks): in source
